progress

+++ src/driver/bk4819.h
@@ -175,7 +175,7 @@ uint8_t BK4819_GetSNR(void);
 uint16_t BK4819_GetVoiceAmplitude(void);
 uint8_t BK4819_GetAfTxRx(void);
 uint8_t BK4819_GetSignalPower(void);
-int32_t BK4819_GetAFCValue();
+int16_t BK4819_GetAFCValue();
 
 bool BK4819_GetFrequencyScanResult(uint32_t *pFrequency);
 BK4819_CssScanResult_t BK4819_GetCxCSSScanResult(uint32_t *pCdcssFreq,
